gcp-hackathon-demos/vision/requirements.txt
Flask==0.12.2
google-cloud-vision==0.27.0
google-cloud-storage==1.4.0
google-cloud-datastore==1.3.0
google-cloud-translate==1.2.0
gunicorn==19.7.1